Implement validator path preference
By choosing only one validator for each subpath (in all validator paths)
found we simply implement a way of overriding existing validators.

Also add a Test for this new feature.  This test is by default ignored
because it has to be ran by itself only since it pollutes the
environment.  It can be run either by:

    cargo test -- --ignored test_override

or

    cargo test -- --test-threads=1 --include-ignored

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>